change the value definition sources
