+++ ui/src/pages/Home/Home.jsx
@@ -1,7 +1,18 @@
+import { useState } from 'react'
+
 // COMPONENTS
 import Chart from './Chart/Chart'
 import Filters from './Filters/Filters'
 import Map from './Map/Map'
+
+// CONSTANTS
+import { 
+  dummyDeviceList,
+  dummyFormList, 
+} from './homeConstants'
+
+// DATE AND TIME
+import moment from 'moment'
 
 // MUIS
 import Divider from '@mui/material/Divider'
@@ -13,6 +24,15 @@ import useStyles from './homeUseStyles'
 
 const Home = () => {
   const classes = useStyles()
+
+  const initialFIlterParameters = {
+    form: dummyFormList[0].text,
+    device: dummyDeviceList[0].text,
+    startTime: moment().subtract(1, 'month').toDate(),
+    endTime: moment().endOf('month').toDate(), 
+  }
+
+  const [ filterParameters, setFilterParameters ] = useState(initialFIlterParameters)
 
   return (
     <Stack className={classes.root}>
@@ -27,7 +47,13 @@ const Home = () => {
       <Divider className={classes.divider}/>
 
       {/* FILTERS */}
-      <Filters/>
+      <Filters
+        initialFIlterParameters={initialFIlterParameters}
+        filterParameters={filterParameters}
+        setFilterParameters={setFilterParameters}
+        formList={dummyFormList}
+        deviceList={dummyDeviceList}
+      />
 
       <Divider className={classes.divider}/>
